feat: accessToken 불러오는 함수 분리
