+++ apis/auth.ts
@@ -9,6 +9,8 @@ import {
   ILoginFormData,
   ILoginResponse,
 } from 'types';
+import LocalStorage from 'utils/storage';
+import { ITokenSet, IRetryAxiosInstanceConfig } from 'types/auth';
 
 interface HeaderProperties extends HeadersDefaults {
   Authorization: string;
@@ -76,4 +78,31 @@ export async function signup({
       message: '서버에러',
     };
   }
+}
+
+export async function getNewAccessToken() {
+  try {
+    const accessToken = await instance
+      .post<IApiResponse<ITokenSet>>('/fake/token', {
+        refreshToken: LocalStorage.getItem<string>('refreshToken', ''),
+      })
+      .then((response) => response.data.data.accessToken);
+    return accessToken;
+  } catch (error) {
+    return Promise.reject(error);
+  }
+}
+
+export function getHeadersWithAuthorizationToken(
+  _config: IRetryAxiosInstanceConfig,
+) {
+  const config = { ..._config };
+  if (!config.headers) {
+    config.headers = {};
+  }
+  config.headers.Authorization = `Bearer ${LocalStorage.getItem<string>(
+    'accessToken',
+    '',
+  )}`;
+  return config;
 }
